Implemented read-only containers
A read-only container does not change its size.
hence, Entries cannot be removed or added by the user.

The question arises then, how to even insert entries into the container.
There is a chain of problems, that result in no entries whatsoever:
 - Default Entries are added by the constructor of the object
 - The container is read from memento
 - Since these entries do not exists in the memento (which is empty be default), they are deleted

A way to overcome this, is by giving the entries fixed names instead of generated from UUID.
With a fixed name, the entries from the memento are in fact restored.